requirements.txt
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
moviepy==1.0.3
numpy==1.19.2
Pillow==7.2.0
proglog==0.1.9
pytube3 @ git+https://github.com/nficano/pytube.git@0d6d526c89615b39f61a9341fb395962e1fbbf68
requests==2.24.0
six==1.15.0
spotipy==2.15.0
tqdm==4.48.2
typing-extensions==3.7.4.3
urllib3==1.25.10
youtube-search==1.1.0
